Reorganized weapons
