+++ M107/config.cpp
@@ -16,13 +16,13 @@ class XtdGearModels
     class JCA_colorBase;
     class CfgWeapons
     {
-        #include "\JCA_AIO_aceax\models\m107Models.hpp"
+        #include "m107Models.hpp"
     };
 };
 class XtdGearInfos
 {
     class CfgWeapons
     {
-        #include "\JCA_AIO_aceax\infos\m107Infos.hpp"
+        #include "m107Infos.hpp"
     };
 };
