+++ M16A4/m16Infos.hpp
