+++ USMC/usmcWepInfos.hpp
