+++ USMC/usmcWepModels.hpp
