+++ mk23/config.cpp
@@ -17,13 +17,13 @@ class XtdGearModels
     class JCA_colorBase;
     class CfgWeapons
     {
-        #include "\JCA_AIO_aceax\models\mk23Models.hpp"
+        #include "mk23Models.hpp"
     };
 };
 class XtdGearInfos
 {
     class CfgWeapons
     {
-        #include "\JCA_AIO_aceax\infos\mk23Infos.hpp"
+        #include "mk23Infos.hpp"
     };
 };
