+++ p226/p226Infos.hpp
